NXDRIVE-1574: DAO has no lock when Drive attempts to restore a db backup
* Moved ensure_no_exception() to the tests/ root.
