+++ tests/__init__.py
@@ -3,6 +3,7 @@ import os
 import os.path
 import shutil
 import sys
+from contextlib import contextmanager
 from typing import Any
 
 import nuxeo.constants
@@ -99,3 +100,31 @@ def setup_sentry() -> None:
 
 
 setup_sentry()
+
+
+@contextmanager
+def ensure_no_exception():
+    """
+    Helper to use as a context manager to check a snippet does not throw any exception.
+    Usefull when one exception is only loggued and not forwared to the parent thread.
+        >>> with ensure_no_exception():
+        ...     # some code where you do not want any exception
+    """
+
+    def error(type_, value, traceback) -> None:
+        """ Install an exception hook to catch any error. """
+        nonlocal received
+        received = True
+        print(type_)
+        print(value)
+        print(repr(traceback))
+
+    received = False
+    excepthook, sys.excepthook = sys.excepthook, error
+
+    try:
+        yield
+    finally:
+        sys.excepthook = excepthook
+
+    assert not received, "Unhandled exception raised!"
