+++ tests/old_functional/__init__.py
@@ -1,8 +1,6 @@
 # coding: utf-8
 import os
-import sys
 import tempfile
-from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, Dict, List, Tuple, Union
 
@@ -55,34 +53,6 @@ def patch_nxdrive_objects():
 
 
 patch_nxdrive_objects()
-
-
-@contextmanager
-def ensure_no_exception():
-    """
-    Helper to use as a context manager to check a snippet does not throw any exception.
-    Usefull when one exception is only loggued and not forwared to the parent thread.
-        >>> with ensure_no_exception():
-        ...     # some code where you do not want any exception
-    """
-
-    def error(type_, value, traceback) -> None:
-        """ Install an exception hook to catch any error. """
-        nonlocal received
-        received = True
-        print(type_)
-        print(value)
-        print(repr(traceback))
-
-    received = False
-    excepthook, sys.excepthook = sys.excepthook, error
-
-    try:
-        yield
-    finally:
-        sys.excepthook = excepthook
-
-    assert not received, "Unhandled exception raised!"
 
 
 def make_tmp_file(folder: Path, content: bytes) -> Path:
